pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hummingbird"
version = "0.1.0"
description = "Flask + React application"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "pandas>=2.0.0",
    "plotly>=5.14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

